requirements.txt
asgiref==3.4.1
boto3==1.18.58
botocore==1.21.58
Django==3.2.7
django-cors-headers==3.9.0
django-storages==1.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
